+++ Changes.md
@@ -1,3 +1,5 @@
+* #668 support adding a team that's not in the schedule and still being able to edit their subjective scores
+
 Release 13.1.0
 ==============